Refactor MeteringPointCreatedEvent + add resolution
