+++ source/dotnet/IntegrationEventListener/MeteringPoints/MeteringPointType.cs
@@ -12,17 +12,35 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Application.MeteringPoints
+namespace Energinet.DataHub.Wholesale.IntegrationEventListener.MeteringPoints
 {
     /// <summary>
     /// WARNING !
     /// Do not change the numeric values since they are used further downstream for aggregation purposes.
     /// </summary>
-    public enum SettlementMethod
+    public enum MeteringPointType
     {
         Unknown = 0,
-        Profiled = 1,
-        NonProfiled = 2,
-        Flex = 3,
+        Consumption = 1,
+        Production = 2,
+        Exchange = 3,
+        VeProduction = 4,
+        Analysis = 5,
+        SurplusProductionGroup = 6,
+        NetProduction = 7,
+        SupplyToGrid = 8,
+        ConsumptionFromGrid = 9,
+        WholesaleService = 10,
+        OwnProduction = 11,
+        NetFromGrid = 12,
+        NetToGrid = 13,
+        TotalConsumption = 14,
+        GridLossCorrection = 15,
+        ElectricalHeating = 16,
+        NetConsumption = 17,
+        OtherConsumption = 18,
+        OtherProduction = 19,
+        ExchangeReactiveEnergy = 20,
+        InternalUse = 21,
     }
 }
