+++ source/dotnet/Tests/Application/MeteringPoints/SettlementMethodTests.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Application.MeteringPoints;
+using Energinet.DataHub.Wholesale.IntegrationEventListener.MeteringPoints;
 using FluentAssertions;
 using Xunit;
